rename gsoc package to dependencyConnections
