+++ src/main/java/org/terasology/rendering/dag/nodes/DownSamplerForExposureNode.java
@@ -20,7 +20,6 @@ import org.terasology.config.RenderingConfig;
 import org.terasology.context.Context;
 import org.terasology.engine.SimpleUri;
 import org.terasology.naming.Name;
-import org.terasology.rendering.dag.gsoc.FboConnection;
 import org.terasology.rendering.opengl.BaseFboManager;
 import org.terasology.rendering.opengl.FBO;
 import org.terasology.rendering.opengl.FboConfig;
